// types.ts
export interface Movie {
  id: number;
  title: string;
  name?: string;
  poster_path: string | null;
  backdrop_path: string | null;
  overview: string;
  release_date?: string;
  first_air_date?: string;
  vote_average: number;
  vote_count: number;
  popularity?: number;
  genres?: { id: number; name: string }[];
  runtime?: number;
  media_type?: 'movie' | 'tv';
  itemType?: 'movie' | 'tv';
  content_ratings?: { results: { iso_3166_1: string; rating: string }[] };
  credits?: { cast: { name:string }[] };
  production_companies?: { name: string }[];
  status?: string;
  original_title?: string;
  original_name?: string;
  original_language?: string;
  recommendations?: { results: Movie[] };
  seasons?: Season[];
  number_of_seasons?: number;
  number_of_episodes?: number;
  currentTime?: number;
  duration?: number;
  // FIX: Add optional videos property based on API response for `append_to_response`.
  videos?: {
    results: {
      key: string;
      site: string;
      type: string;
    }[];
  };
  images?: {
    logos: {
      file_path: string;
      iso_639_1: string;
    }[];
  };
}
// ...
export interface Season {
  id: number;
  season_number: number;
  name: string;
  poster_path: string | null;
  episode_count: number;
}
// ...
export interface ChatMessage {
  id: number;
  user: {
    name: string;
    avatar: string;
  };
  text: string;
  isJoin?: boolean;
}